requirements.txt
requests
fastapi
uvicorn[standard]
openai
python-dotenv
typer[all]
sphinx